refactor(trip-form): move mode selector story into its own file
